requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.2
pandas==2.1.4
pymongo==4.6.0
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==1.4.50
